feat: init new sections ui homepage
